MLOps/pyproject.toml
[tool.poetry]
name = "mlops"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
torch = "^2.1.0"
torchvision = "^0.16.0"
pandas = "^2.1.1"
pytorch-lightning = "^2.1.2"
gitpython = "^3.1.40"
dvc = "^3.30.3"
dvc-gdrive = "^2.20.0"
fire = "^0.5.0"
hydra-core = "^1.3.2"
mlflow = "^2.8.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
target-version = ["py37"]
skip-string-normalization = true

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 79
lines_after_imports = 2
